Add DHCP support for remote syslog
Reviewed By: dustin

Add the option to log to remote syslog servers.  The syslog servers
are specified via DHCP response option 7, which is the 'log-servers'
option in ISC DHPC.

Tested using ISC DHCP server option 'log-servers', using multiple
remote syslog servers.

Also tested against syslog-ng and rsyslogd servers.

Also tested changing the DHPC configure in the middle of the discovery
process.  The system reacted as expected, killing the old syslogd
process and starting a new one with the updated remote syslog parameters.

+++ rootconf/default/etc/init.d/syslogd.sh
@@ -1,26 +1,42 @@
 #!/bin/sh
 
 cmd="$1"
-remote="$2"
 
 daemon="syslogd"
 
-ARGS="-b 3 -D"
-[ -n "$remote" ] && ARGS="$ARGS -L -R $remote"
+ARGS="-b 3 -D -L"
 
 . /scripts/functions
 
+ARG_FILE="${ONIE_RUN_DIR}/syslogd.args"
+LOGSRVS=$(cat "${ONIE_RUN_DIR}/dhcp.logsrv")
+for r in $LOGSRVS ; do
+    ARGS="$ARGS -R $r"
+done
+
+# Only restart the syslogd if the args have changed
+OLD_ARGS=
+[ -r "$ARG_FILE" ] && OLD_ARGS=$(cat "$ARG_FILE")
+
 case $cmd in
-    start)
+    start|discover)
+        if [ "$OLD_ARGS" = "$ARGS" ] && pidof -s $daemon > /dev/null 2>&1 ; then
+            # The args did not change and the server is still running.
+            # Do nothing.
+            return 0
+        fi
+        # kill the old one
         killall $daemon > /dev/null 2>&1
         $daemon $ARGS
+        echo "$ARGS" > $ARG_FILE
         ;;
+
     stop)
         log_begin_msg "Stopping: $daemon"
         killall $daemon > /dev/null 2>&1
         log_end_msg
         ;;
+
     *)
         
 esac
-
